Auto-Generation Complete
Completed the Auto-Generation of the Manga List.

All the user has to do is supply a list of their Manga Hyperlinks via the "mangaList.js" file and site will convert that into a list of clickable links, with the names stripped out and with Block Open buttons.

Allowing the user to open 10 links at a time.

*Note*: This can easily be changed by changing the variable in the links file.